+++ app/src/main/java/com/vultisig/wallet/data/usecases/UseCasesModule.kt
@@ -52,4 +52,10 @@ internal interface UseCasesModule {
         impl: DecomressQrUseCaseImpl
     ): DecompressQrUseCase
 
+    @Binds
+    @Singleton
+    fun bindEnableTokenUseCase(
+        impl: EnableTokenUseCaseImpl
+    ): EnableTokenUseCase
+
 }
